refactor: DTO 를 사용하는 계층으로 이동
