+++ src/main/java/wooteco/subway/controller/ControllerAdvice.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.dto.response.ErrorResponse;
+import wooteco.subway.controller.dto.ErrorResponse;
 
 @RestControllerAdvice
 public class ControllerAdvice {
