+++ src/main/java/wooteco/subway/controller/SectionController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import wooteco.subway.dto.request.SectionRequest;
+import wooteco.subway.service.dto.request.SectionRequest;
 import wooteco.subway.service.SectionService;
 
 import javax.validation.Valid;
