+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -6,7 +6,7 @@ import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
-import wooteco.subway.dto.request.SectionRequest;
+import wooteco.subway.service.dto.request.SectionRequest;
 
 @Service
 @Transactional(readOnly = true)
