+++ src/main/java/wooteco/subway/service/dto/request/SectionRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.request;
+package wooteco.subway.service.dto.request;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
