+++ src/main/java/wooteco/subway/service/dto/response/LineResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto.response;
+package wooteco.subway.service.dto.response;
 
 import java.util.List;
 import wooteco.subway.domain.Line;
