+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.response.StationResponse;
+import wooteco.subway.service.dto.response.StationResponse;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
